+++ tests/testthat/test-helpers-shared.R
@@ -5,10 +5,10 @@ context("Shared: misc. functions")
 test_that("suggested_package throws correct errors", {
   expect_error(suggested_package("NOPACKAGE"),
                "Please install the NOPACKAGE package")
-  expect_error(suggested_package(c("testthat", "NOPACKAGE")),
-               "Please install the NOPACKAGE package")
+  expect_error(suggested_package(c("testthat", "gridExtra")), "length")
   expect_silent(suggested_package("testthat"))
-  expect_silent(suggested_package(c("testthat", "gridExtra")))
+  expect_silent(suggested_package("testthat", min_version = "0.0.1"))
+  expect_error(suggested_package("testthat", min_version = "100000.0.0"))
 })