+++ R/mcmc-diagnostics.R
@@ -216,6 +216,17 @@ mcmc_neff <- function(ratio, ..., size = NULL) {
   check_ignored_arguments(...)
   data <- mcmc_neff_data(ratio)
 
+  max_ratio <- max(ratio, na.rm = TRUE)
+  if(max_ratio < 1.25) {
+    additional_breaks <- numeric(0)
+  } else if(max_ratio < 1.5) {
+    additional_breaks <- 1.25
+    additional_labels <- "1.25"
+  } else {
+    additional_breaks <- seq(1.5, max_ratio, by = 0.5)
+  }
+  breaks <- c(0, 0.1, 0.25, 0.5, 0.75, 1, additional_breaks)
+
   ggplot(
     data,
     mapping = aes_(
@@ -236,9 +247,9 @@ mcmc_neff <- function(ratio, ..., size = NULL) {
     scale_fill_diagnostic("neff") +
     scale_color_diagnostic("neff") +
     scale_x_continuous(
-      breaks = c(0, 0.1, 0.25, 0.5, 0.75, 1),
-      labels = c("0", "0.1", "0.25", "0.5", "0.75", "1"),
-      limits = c(0, 1.05),
+      breaks = breaks,
+      labels = as.character(breaks), #as.character truncates trailing zeroes, while ggplot default does not
+      limits = c(0, max(1, max_ratio) + 0.05),
       expand = c(0, 0)) +
     yaxis_text(FALSE) +
     yaxis_title(FALSE) +
@@ -629,8 +640,8 @@ new_neff_ratio <- function(x) {
 
 validate_neff_ratio <- function(x) {
   stopifnot(is.numeric(x), !is.list(x), !is.array(x))
-  if (any(x < 0 | x > 1, na.rm = TRUE)) {
-    stop("All neff ratios must be between 0 and 1.", call. = FALSE)
+  if (any(x < 0, na.rm = TRUE)) {
+    stop("All neff ratios must be positive.", call. = FALSE)
   }
   x
 }