+++ NEWS.md
@@ -3,8 +3,13 @@
 (GitHub issue/PR numbers in parentheses)
 
 * The [LOO predictive
-checks](http://mc-stan.org/bayesplot/reference/PPC-loo.html) now require __loo__
-package version >= 2.0.0. (#139)
+checks](http://mc-stan.org/bayesplot/reference/PPC-loo.html) now require 
+**loo** version `>= 2.0.0`. (#139)
+
+* `mcmc_pairs()` now has an argument `grid_args` to provide a way of passing
+optional arguments to `gridExtra::arrangeGrob()`. This can be used to add a
+title to the plot, for example. (#143)
+
 
 # bayesplot 1.5.0