+++ NEWS.md
@@ -2,9 +2,15 @@
 
 (GitHub issue/PR numbers in parentheses)
 
+* The [LOO predictive
+checks](http://mc-stan.org/bayesplot/reference/PPC-loo.html) now require 
+**loo** version `>= 2.0.0`. (#139)
+
 * `mcmc_pairs()` now has an argument `grid_args` to provide a way of passing
 optional arguments to `gridExtra::arrangeGrob()`. This can be used to add a
 title to the plot, for example. (#143)
+
+* `ppc_ecdf_overlay()` gains an argument `discrete`, which is `FALSE` by default, but can be used to make the Geom more appropriate for discrete data. (#145)
 
 
 # bayesplot 1.5.0